# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API OmniGest",
        "description": "Documentação da API Omnigest",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/material/{id_material}": {
            "get": {
                "tags": [
                    "Material"
                ],
                "summary": "Obter material por ID",
                "description": "Retorna os detalhes de um material específico",
                "operationId": "getMaterialById",
                "parameters": [
                    {
                        "name": "id_material",
                        "in": "path",
                        "description": "ID do material",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do material",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Material não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Material"
                ],
                "summary": "Atualizar material",
                "description": "Atualiza os dados de um material",
                "operationId": "updateMaterial",
                "parameters": [
                    {
                        "name": "id_material",
                        "in": "path",
                        "description": "ID do material",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_unidade_mte": {
                                        "description": "ID da unidade",
                                        "type": "integer"
                                    },
                                    "des_material_mte": {
                                        "description": "Descrição do material",
                                        "type": "string"
                                    },
                                    "vlr_material_mte": {
                                        "description": "Valor do material",
                                        "type": "integer"
                                    },
                                    "id_centro_custo_mte": {
                                        "description": "ID do centro de custo",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Material atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Material"
                ],
                "summary": "Excluir material",
                "description": "Desativa um material",
                "operationId": "deleteMaterial",
                "parameters": [
                    {
                        "name": "id_material",
                        "in": "path",
                        "description": "ID do material",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Material excluído com sucesso"
                    }
                }
            }
        },
        "/material": {
            "post": {
                "tags": [
                    "Material"
                ],
                "summary": "Criar material",
                "description": "Cria um novo material",
                "operationId": "createMaterial",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_unidade_mte": {
                                        "description": "ID da unidade",
                                        "type": "integer"
                                    },
                                    "des_material_mte": {
                                        "description": "Descrição do material",
                                        "type": "string"
                                    },
                                    "vlr_material_mte": {
                                        "description": "Valor do material",
                                        "type": "integer"
                                    },
                                    "id_centro_custo_mte": {
                                        "description": "ID do centro de custo",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Material criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/{id_user}": {
            "get": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Obtém os dados de um usuário",
                "operationId": "18695312a79f258286f2c79db840375c",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Atualiza os dados de um usuário",
                "operationId": "705b25a58747427421fe2e97d32fdafa",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "url_img_user": {
                                        "type": "string",
                                        "format": "uri"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário atualizado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validação falhou",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuario"
                ],
                "summary": "Deleta um usuário",
                "operationId": "e2eb016fcbed0b84a900ab05baf368e7",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário deletado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuário deletado com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Material": {
                "properties": {
                    "id_material_mte": {
                        "type": "integer"
                    },
                    "id_unidade_mte": {
                        "description": "Chave estrangeira referenciando a unidade",
                        "type": "integer"
                    },
                    "id_empresa_mte": {
                        "description": "Chave estrangeira referenciando a empresa",
                        "type": "integer"
                    },
                    "des_material_mte": {
                        "type": "string"
                    },
                    "vlr_material_mte": {
                        "type": "integer"
                    },
                    "is_ativo_mte": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "id_centro_custo_mte": {
                        "description": "Chave estrangeira referenciando o centro de custo",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "User": {
                "type": "array",
                "items": {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "name": {
                            "type": "string"
                        },
                        "email": {
                            "type": "string"
                        },
                        "url_img_user": {
                            "type": "string",
                            "nullable": true
                        },
                        "created_at": {
                            "type": "string",
                            "format": "date-time"
                        },
                        "updated_at": {
                            "type": "string",
                            "format": "date-time"
                        }
                    },
                    "type": "object"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}